下架更新

+++ pages/course/detail.vue
@@ -86,10 +86,12 @@
 			};
 		},
 		onLoad(opt) {
-			this.getUserIndex();
 			this.id = opt.id;
 			this.getArchiveDetail();
-			this.checkCollection();
+			if (this.vuex_token) {
+				this.getUserIndex();
+				this.checkCollection();
+			}
 		},
 		onShow() {},
 		computed: {
@@ -125,6 +127,18 @@
 		methods: {
 
 			handleCollect() {
+				if (!this.vuex_token) {
+					uni.showToast({
+						title: '请登录',
+						icon: 'none',
+						success() {
+							uni.navigateTo({
+								url: '/pages/login/login'
+							})
+						}
+					})
+					return
+				}
 				if (this.isCollect) {
 					this.$api.deleteCollection({
 						aid: this.id
